example/qualify-email-slack-bot/requirements.txt
email_validator==1.3.0
openai==0.25.0
playwright==1.28.0
python-dotenv==0.21.0
